Create Leetcode 3190
Question.  Find Minimum Operations to Make All Elements Divisible by Three?

Answer. In c++ there we find how many operation we take to make it divisible by 3. 
There is a main concept on this is element is already divisible by 3 we don't need to do any operation .
If element is not divisible by 3 simply we update our count.

At the end the counter (cnt) is our answer.