2551  Put Marbles in Bags.
This is a very good problem...... Hard marked but if you good in dry run and some relation you found this is very easy to solve.....don't think too much... Simply dry run you will found it.

Some mistake i do in dry run.  I don't do properly dry run it.  In the place of simple thinking i am think too advance for this problem...

So this problem in good . I like it.