73 Set Matrix Zeroes
This is the medium problem in this  if we see 0 in anywhere in our matrix simply  set 0 all its column and rows of the matrix .

So We take here two vector array which is zero in row and zero in col.  Which are help us to find the row and col where  the location and where we should put the zero.